Create FavoriteIcon component
