+++ src/Components/App/App.js
@@ -23,11 +23,11 @@ function App() {
     setDrinks(data)
   }
 
-  function toggleFavorite(drink) {
-    if (favorites.find(favorite => favorite.idDrink === drink.idDrink)) {
-      setFavorites(prev => prev.filter(favorite => favorite.idDrink !== drink.idDrink))
+  function toggleFavorite(drinkId) {
+    if (favorites.find(favoriteId => favoriteId === drinkId)) {
+      setFavorites(prev => prev.filter(favoriteId => favoriteId !== drinkId))
     } else {
-      setFavorites(prev => [...prev, drink])
+      setFavorites(prev => [...prev, drinkId])
     }
   }
 
@@ -37,7 +37,8 @@ function App() {
 
   return <Routes>
     <Route path='/' element={<HomePage drinks={drinks} favorites={favorites} toggleFavorite={toggleFavorite} error={error}/>}/>
-    <Route path='/:id' element={<RecipePage  favorites={favorites} toggleFavorite={toggleFavorite} error={error}/>} />
+    <Route path='/:id' element={<RecipePage  favorites={favorites} toggleFavorite={toggleFavorite}/>} />
+    <Route path='*' />
   </Routes>
 }
