+++ src/Components/RecipePage/RecipePage.scss
@@ -1,0 +1,15 @@
+.recipe__img {
+  height: 20rem;
+}
+
+.page {
+  background-color: #fcf9f0;
+  display: flex;
+  padding: 2rem;
+}
+
+.main {
+  display: flex;
+  flex-direction: column;
+  padding: 2rem;
+}
